requirements.txt
# === Core PDF processing ===
PyMuPDF==1.22.0
pytesseract==0.3.10
Pillow==9.5.0

# === Natural Language Processing ===
spacy==3.7.2
scikit-learn==1.3.0
numpy<2.0
srsly==2.4.8
confection==0.1.4
thinc==8.2.2

# === Sentence Transformers (Lightweight Embedding Model) ===
sentence-transformers==2.2.2
torch==2.0.0        # If you want minimal, pure CPU use this version
transformers==4.40.0

# === Utilities ===
ujson==5.9.0
loguru==0.7.2
cachetools==5.3.2

# === CLI Feedback
tqdm==4.66.2
